Move JSHandle to js_handle.go
